backend/deployment_requirements.txt
waitress==3.0.1
whitenoise==6.7.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2-binary==2.9.9
